+++ Model/DashboardRow/ComposerAutoloader.php
@@ -35,30 +35,30 @@ class ComposerAutoloader extends \MageHost\PerformanceDashboard\Model\DashboardR
      */
     public function load()
     {
-        $this->setTitle('Composer autoloader');
+        $this->setTitle(__("Composer autoloader"));
 
         /** @var null|\Composer\Autoload\ClassLoader $classLoader */
         $classLoader = null;
-        foreach (spl_autoload_functions() as $func) {
-            if (is_array($func) &&
-                 $func[0] instanceof \Composer\Autoload\ClassLoader ) {
-                $classLoader = $func[0];
+        foreach (spl_autoload_functions() as $function) {
+            if (is_array($function) &&
+                $function[0] instanceof \Composer\Autoload\ClassLoader ) {
+                $classLoader = $function[0];
                 break;
             }
         }
 
         if (empty($classLoader)) {
             $this->setStatus(self::STATUS_UNKNOWN);
-            $this->setInfo("Could not find Composer AutoLoader.");
+            $this->setInfo(__("Could not find Composer AutoLoader."));
             return;
         }
 
         if (array_key_exists('Magento\Config\Model\Config', $classLoader->getClassMap())) {
             $this->setStatus(self::STATUS_OK);
-            $this->setInfo("Composer's autoloader is optimized");
+            $this->setInfo(__("Composer's autoloader is optimized"));
         } else {
             $this->setStatus(self::STATUS_PROBLEM);
-            $this->setInfo("Composer's autoloader is not optimized.");
+            $this->setInfo(__("Composer's autoloader is not optimized."));
             $this->setAction(__("Execute: 'composer install -o'"));
         }
     }